Add uniq_id solution
See the rules for the preprocessor to understand how this macro
works:
"Macro arguments are completely macro-expanded before they are
substituted into a macro body, unless they are stringized or pasted with
other tokens. After substitution, the entire macro body, including the
substituted arguments, is scanned again for macros to be expanded. The
result is that the arguments are scanned twice to expand macro calls in
them" (https://gcc.gnu.org/onlinedocs/cpp/Argument-Prescan.html#Argument-Prescan).