Implement with a set and map
This implementation is similar to the
course solution but uses set instead of
the custom function.
It seems that I've already implemented
the similar version. It was the the first
version (that had not been pushed to
the git). It was very similar but it
didn't pass the tests. May be because
I didn't move the whole group of strings
but each string in a group individually.