Implement merge_sort_3 (with error but it works)
After examining the coursera's solution and error has been founded.
The function calls MergeSort on the elements of the original vector using
range_begin not on the elements of the local vector.
Don't know how it works but it works. I presume that the local vector is not used
at all (so there is a inplace merging).

To do:
* try to delete local vector "v"; - it doesn't work
(Failed case #1/23: (Wrong answer) Объекты перемещаются неверное количество раз)

* fix the solution according to the coursera's